Fix latest pydub AudioSegment merge crash
When AudioSegment.empty() is used we end up with some empty parameters
which then in turn mean we can't use this variable to += a real audio
content on the latest pydub present on debian testin. To address this just
use AudioSegment.silent() with proper settings.

While at it add a few new cli options to let us debug things easily:

  * --debug flag: Enables verbose logging across the application to assist
                  with troubleshooting.
  * --keep-files: Preserves temporary audio files used in TTS synthesis for
                  inspection. Useful when debugging audio generation or
                  concatenation issues.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>